+++ Suspension/src/Suspension.Core/Operations/RunAnalysis/Fatigue/RectangularProfile/RunRectangularProfileFatigueAnalysis.cs
@@ -1,5 +1,7 @@
 ﻿using MudRunner.Commons.Core.ConstitutiveEquations.Fatigue;
 using MudRunner.Suspension.Core.Operations.RunAnalysis.Static.RectangularProfile;
+using MudRunner.Suspension.DataContracts.RunAnalysis.Fatigue;
+using System.Threading.Tasks;
 using DataContract = MudRunner.Commons.DataContracts.Models.Profiles;
 
 namespace MudRunner.Suspension.Core.Operations.RunAnalysis.Fatigue.RectangularProfile
@@ -18,5 +20,18 @@ namespace MudRunner.Suspension.Core.Operations.RunAnalysis.Fatigue.RectangularPr
             IRunRectangularProfileStaticAnalysis runStaticAnalysis,
             IFatigue<DataContract.RectangularProfile> fatigue) : base(runStaticAnalysis, fatigue)
         { }
+
+        /// <summary>
+        /// Asynchronously, this method validates the <see cref="RunFatigueAnalysisRequest{RectangularProfile}"/>.
+        /// </summary>
+        /// <param name="request"></param>
+        /// <returns></returns>
+        protected override Task<RunFatigueAnalysisResponse> ValidateOperationAsync(RunFatigueAnalysisRequest<DataContract.RectangularProfile> request)
+        {
+            RunFatigueAnalysisResponse response = new();
+            response.SetSuccessOk();
+
+            return Task.FromResult(response);
+        }
     }
 }