+++ Suspension/src/Suspension.DataContracts/RunAnalysis/Dynamic/RunDynamicAnalysisResponse.cs
@@ -1,0 +1,9 @@
+﻿using MudRunner.Commons.DataContracts.Operation;
+
+namespace MudRunner.Suspension.DataContracts.RunAnalysis.Dynamic
+{
+    /// <summary>
+    /// It represents the response content of RunDynamicAnalysis operation.
+    /// </summary>
+    public class RunDynamicAnalysisResponse : OperationResponseBase<RunDynamicAnalysisResponseData> { }
+}